pplied a path provided by Mark Rosenstein <mar@datwin.inergy.com>. His comments on the patch are below:
The patches below make two changes to the 2.0.8 module.

There's a bug in the code which garbage collects the FastCgiServerInfo
structure after a script dies.  It garbage collects the structure, but
leaves the linked list pointed at the freed memory, with no reference
to the rest of the chain.  I suspect that this was causing some of the
obscure server bugs and ASSERTion errors I was getting.  Our scripts
do exit with some regularity when certain conditions occur.

I've also added a new configuration option.  FCGIConfig now accepts an
option "-restart".  If present, dynamic FCGI scripts will be restarted
when they die.  We use this because our application consists of about
50 scripts; restarting the server caused a huge load spike when all of
these were started at once.  This way, they are started dynamically as
they are needed, and will be restarted if they exit.