3 further patches by Mark Rosenstein, mar@datwin.inergy.com.
The patch below contains three fixes, all related to the "-restart"
switch.  Roy, I don't think any of these will solve your problem, but
it certainly makes my system more stable.

The bug I found in the current code is that if the timer goes off to
killDynamicApps between one dying and getting restarted, it could
leave an inconsistent structure and fail to ever restart the app.

It keeps better statistics about dynamic apps, incrementing
numFailures and numRestarts.  It also backs off on the speed of
successive restarts until it's waiting 10 seconds between restarts.
This is useful in the case where you accidentally put in a faulty
script, so that it doesn't just loop as fast as it can.

It also avoids a potential divide-by-zero error while calculating
statistics which may result in the process manager crashing.  I think
that this could only happen if the first bug mentioned above occurs,
but it's safer to just make sure that the divide-by-zero cannot
happen.
					-Mark