Auto-restart of changed binaries patch by Mark Rosenstein <mar@datwin.inergy.com>.
OK, here's the patch to get it to automatically restart a dynamic
application process if the binary is updated on disk.  If
FCGI_DEFAULT_AUTOUPDATE is set to 1, or "-autoUpdate" flag is given in
the FCGIconfig line, then this feature is active.  Because we only use
dynamic processes here, I haven't gotten this to work with declared
AppClasses.
					-Mark