Includes a Stronghold patch by Mark Rosenstein <mar@inergy.com>
The latest version of fastcgi (2.0.10), when compiled under the latest
stronghold (2.2), does not initialize correctly.  It looks like under
stronghold, the first call to ModFastCgiInit happens before we know
what uid and gid we should be running under, so the permissions on the
fastcgi dynamic directory and mbox file are wrong.  The following
patch works around this by changing the ownership later.
			   -Mark