Based on the patch provided by Alan Rykhus <alan.rykhus@mankato.msus.edu>...
In working with the code for something else I came across what I think 
is an error that may be related to this problem. It only effects things 
if you are running fastcgi statically.

In the mod_fastcgi module, version 13, in mod_fastcgi.c in the the 
FastCgiProcMgr function at line 3233. The manager is starting up new 
copies of each type of fastcgi script. The line is:

	s->numProcesses++;

If it is running dynamically it should do this to keep track of how 
many processes are running. But if you are running a static number of 
processes it increases the number of processes that it was told to 
start in the AppClass directive in your .conf file. For instance, if 
you had -processes 3, the manager thinks it was set to 6 when its done 
firing off all of the scripts. I think this line should be wrapped 
with:

	if (s->directive == APP_CLASS_DYNAMIC) {
		s->numProcesses++;
	}

I've made this change to our code and we've had a couple of less 
problems.

al

----------------------
Alan Rykhus
MnSCU/PALS (507)389-1975
alan.rykhus@mankato.msus.edu