Handle SIGUSR1 and SIGHUP same as SIGTERM.  Apache  sends to its process group SIGUSR1 and SIGHUP when  received (soft and hard restart).  These signals were  previously being ignored and thus the ProcMgr likely  had to wait for the sleepSeconds (default=300) alarm  to expire in order return from sigsuspend() and  notice that the KillMgr is no longer present and  begin to shutdown.  Apache isn't too patient, if each  of its children aren't collected it will eventually  send SIGKILL - which, in our case, the KillMgr will  receive.  The KillMgr is sacrificed, but the ProcMgr  can continue to shutdown gracefullly.
Of course this means you shouldn't send SIGHUP or
SIGUSR1 (or SIGTERM) to the ProcMgr unless you want it
to shutdown and thats not a good idea if Apache is
still running.

If your fcgi apps catch SIGTERM and take awhile to
exit and you use AppClass, sending SIGUSR1 or SIGHUP
to apache may result in a failed restart - this means
you no longer have a server.  If you periodically send
SIGUSR1 or SIGHUP you should check that apache is
running.

This was tested only against Apache 1.2.6. Apache
1.3.x uses a better approach to killing off its
children, but I haven't tested it (and can't remember
the diffrence in approach).	  We may be able to do
something slicker for 1.3, speak up if have ideas.