- A fix and enhancement submitted by Paul G. Weiss [PGWeiss@arity.com]..
(1) It fixes the problem that occurs when the process
manager is repeated interrupted by fastcgi process
terminations, thereby never getting to run RemoveRecords or
KillDynamicProcs.  The fix is to record the return value of
the alarm(0) call and use it to determine how long to sleep
next.

(2) Allow a -listen-queue-depth argument to FCGIConfig just
like in AppClass. Use this value in creating the socket for
the dynamic apps.  I found that I was getting a lot of
errors when the listen queue depth was set to the default
value (which is 5).

- I added support for the -appConnTimeout arg in AppClass.
Previously, static apps used the same connect() timeout
supplied in the FCGIConfig directive (or the default of 15
seconds).

- Note: The documentation has not been updated to include
either of these two enhancements.

- Cast the 4th arg of getsockopt() to avoid a compiler
warning (on Solaris?).

- Change the sprintf() format string for (pid_t)procMgr
from "%ld" to "%d" to avoid a compiler warning (gcc/Linux).
I don't think this will break anything, but maybe a cast
would be better?