Added suexec support based on patches provided by Brian Grossman <brian@SoftHome.net>.
To enable it you have to enable suexec for CGI.

Using suexec improperly (for CGI and/or FCGI) can pose SERIOUS SECURITY
problems, see the suexec discussion in the Apache documentation.

It works nearly identically to suexec under	CGI.  Because FCGI
applications are persistant, a check has to be done within the module's
request handler to be sure the fcgi application associated with the
socket used for a connect() is running as the uid/gid of the Virtual
Server servicing the request.

There's still a few items I'd like to revisit - there's notes in the
code.