- Note: docs need updating.
- Changed the default directory for the Unix/Domain sockets from /tmp to
/tmp/fcgi.

- Greatly improved suexec support.  No restrictions are imposed based on
the permissions of the FastCGI server.  A new directive determines its
functionality.  It can be disabled even if its enabled within Apache.  A
different suexec program can be specified than that used for Apache CGI.
FastCGI servers are shared across Apache servers, but access is limited
based on the uid/gid of the Apache server handling the request and the
uid/gid of the FastCGI server.  If a FastCGI server is running with a
different uid/gid than that needed to handle a request to that server, a
new dynamic instance of the FastCGI server is started running with the
needed uid/gid (take care not to allow this to occur if your suexec
restricts its invocation to one uid/gid pair or you'll end up with a ton
of failed attempts to exec the program - the stock Apache suexec does
this).

- All memory allocation and file descripttors are now handled via Apache
pools.  This will allow the use of the Apache hard_timeout rather than
the soft_timeout in an upcoming patch.

- Fixed a memory leak in GetHashedPath().

- Sockets used by the FastCGI servers are now not opened until afer the
PM forks from the Apache server.  This eliminates a pile of unnecessary
open FDs in all of the Apache processes.  It also allows the apachectl
script to be used with the configtest arg.  Apache should also handle
restart and graceful much better now.

- The dependency on the TCL string routines have been eliminated.

- The OS abstaction (that wasn't) was elminated.

- All of the configuration and socket setup was rewritten.

- Misc. unused variables were removed.

- Probably some other misc stuff...

Modified Files: Makefile.tmpl fcgibuf.c fcgibuf.h mod_fastcgi.c 
Removed Files:  Makefile.sample fcgios.c fcgios.h fcgitcl.c fcgitcl.h