Fastcgi module for Apache 2.2
OK, after some investigation I found out that the "new" (apr_*) symbols
were introduced beginning with 2.0 and that the older symbols were
defines were dropped.

Therefore I believe the simplest/shortest way to make mod_fastcgi work
with 1.3, 2.0 and 2.2 is to reintroduce these defines ourselves when
necessary. The attached patch does this and additionaly fixes a problem
in Makefile.AP2 under 2.2 (still works under 2.0 though).

It was tested on 1.3.33, 2.0.55 and 2.2 and compiles and loads on all
three versions. I believe it is suitable for inclusion in the official
code base.